Rewrite v3 time-dependent tests with testing/synctest
Use build tags to split expiry tests into two variants:
- cache_expire_test.go (!go1.25): original time.Sleep-based tests
- cache_expire_synctest_test.go (go1.25): deterministic synctest-based tests

Bump v3/go.mod to go 1.23 (required for sync timer channels)
and CI to Go 1.25 (where synctest is stable). v1 and v2 remain
unchanged on go 1.20.
