+++ .github/workflows/ci-v3.yml
@@ -23,7 +23,7 @@ jobs:
       - name: set up go
         uses: actions/setup-go@v5
         with:
-          go-version: "1.20"
+          go-version: "1.25"
           cache-dependency-path: v3/go.sum
         id: go
